move fetchNamespaces
